move ping to netcheck, since it's only used there. This, to further reduce crate deps
